Refactor DER encoding helpers.
